+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/CityTablePanel.java
@@ -99,7 +99,7 @@ final class CityTablePanel extends ChartTablePanel {
 	private static FilterTableCellRenderer nameRenderer(CityTableModel tableModel) {
 		return EntityTableCellRenderer.builder(City.NAME, tableModel)
 						.foreground((table, row, column, value) -> {
-							Entity city = (Entity) table.model().items().visible().itemAt(row);
+							Entity city = table.model().items().visible().itemAt(row);
 							if (Objects.equals(city.get(City.ID), city.get(City.COUNTRY_FK).get(Country.CAPITAL))) {
 								return Color.GREEN;
 							}